+++ tmx/object.go
@@ -8,18 +8,41 @@ import (
 
 // Object structure: https://doc.mapeditor.org/en/stable/reference/tmx-map-format/#object
 type Object struct {
-	XMLName  xml.Name `xml:"object"`
-	ID       int      `xml:"id,attr"`       // Unique ID of the object. Each object that is placed on a map gets a unique id. Even if an object was deleted, no object gets the same ID. Can not be changed in Tiled. (since Tiled 0.11)
-	Name     string   `xml:"name,attr"`     //  The name of the object. An arbitrary string (defaults to “”).
-	Type     string   `xml:"type,attr"`     //  The type of the object. An arbitrary string (defaults to “”).
-	X        float64  `xml:"x,attr"`        //  The x coordinate of the object in pixels.
-	Y        float64  `xml:"y,attr"`        //  The y coordinate of the object in pixels.
-	Width    float64  `xml:"width,attr"`    //  The width of the object in pixels (defaults to 0).
-	Height   float64  `xml:"height,attr"`   //  The height of the object in pixels (defaults to 0).
-	Rotation float64  `xml:"rotation,attr"` //  The rotation of the object in degrees clockwise around (x, y) (defaults to 0).
-	GID      int      `xml:"gid,attr"`      //  A reference to a tile (optional).
-	Visible  bool     `xml:"visible,attr"`  //  Whether the object is shown (1) or hidden (0). Defaults to 1.
-	Template string   `xml:"template,attr"` //  A reference to a template file (optional).
+	XMLName xml.Name `xml:"object"`
+
+	// Unique ID of the object. Each object that is placed on a map gets a unique id. Even if an object was deleted, no
+	// object gets the same ID. Can not be changed in Tiled. (since Tiled 0.11)
+	ID int `xml:"id,attr"`
+
+	// The name of the object. An arbitrary string (defaults to “”).
+	Name string `xml:"name,attr"`
+
+	// The type of the object. An arbitrary string (defaults to “”).
+	Type string `xml:"type,attr"`
+
+	// The x coordinate of the object in pixels.
+	X float32 `xml:"x,attr"`
+
+	// The y coordinate of the object in pixels.
+	Y float32 `xml:"y,attr"`
+
+	// The width of the object in pixels (defaults to 0).
+	Width float32 `xml:"width,attr"`
+
+	// The height of the object in pixels (defaults to 0).
+	Height float32 `xml:"height,attr"`
+
+	// The rotation of the object in degrees clockwise around (x, y) (defaults to 0).
+	Rotation float32 `xml:"rotation,attr"`
+
+	// A reference to a tile (optional).
+	GID int `xml:"gid,attr"`
+
+	// Whether the object is shown (1) or hidden (0). Defaults to 1.
+	Visible bool `xml:"visible,attr"`
+
+	// A reference to a template file (optional).
+	Template string `xml:"template,attr"`
 
 	// While tile layers are very suitable for anything repetitive aligned to the tile grid, sometimes you want to
 	// annotate your map with other information, not necessarily aligned to the grid. Hence the objects have their
